使用新搜索功能
