+++ app/components/nav/SearchButton.vue
@@ -1,11 +1,11 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+const { toggleSearchStatus } = useLayoutStatus()
+</script>
 
 <template>
-  <NuxtLink to="/search">
-    <mdui-button-icon>
-      <mdui-icon-search></mdui-icon-search>
-    </mdui-button-icon>
-  </NuxtLink>
+  <mdui-button-icon @click="toggleSearchStatus()">
+    <mdui-icon-search></mdui-icon-search>
+  </mdui-button-icon>
 </template>
 
 <style scoped></style>
