+++ app/composables/layout-status.ts
@@ -1,0 +1,17 @@
+const [drawerStatus, toggleDrawerStatus] = useToggle(false)
+const [searchStatus, toggleSearchStatus] = useToggle(false)
+
+const closeAll = () => {
+  drawerStatus.value = false
+  searchStatus.value = false
+}
+
+export const useLayoutStatus = () => {
+  return {
+    drawerStatus,
+    toggleDrawerStatus,
+    searchStatus,
+    toggleSearchStatus,
+    closeAll,
+  }
+}
