+++ tsconfig.json
@@ -4,7 +4,8 @@
   "compilerOptions": {
     "types": [],
     "paths": {
-      "/*": ["./*"]
+      "/*": ["./*"],
+      "~/*": ["./app/*"],
     }
   }
 }
